Use ${PREFIX}_DEVELOPMENT_FORCE_CELERY to force task wrapper to return a celery worker
With no celery worker deployed, Metadash will return a dumb task that simulates celery API but execute the task within main thread. This helps sometimes, for example when development locally. However, when huge tasks are queued and executed, Metadash will timeout without a worker. This switch enforces the tasks to be dispatched to a worker and ensures the main thread does not timeout.